todoproject/requirements.txt
django
djangorestframework
drf_yasg
djangorestframework_simplejwt
bcrypt